+++ ftree/types.h
@@ -3,9 +3,9 @@
 
 typedef struct file* FILES;
 struct file {
-  unsigned int iblock;
+  size_t iblock;
   const char name[11];
   const time_t create_date;
   time_t last_access;
-  unsigned int bytes;
+  size_t bytes;
 };
